feat: 로또 결과 객체 생성
